Unify "pressure source sample" and "pending sample descriptor" concepts
They were basically the same: a struct with data and a timestamp. In the
"pending sample descriptor" case, the data was always a PressureState,
whereas in the former's case it was an implementation-defined value.

Since a virtual pressure source is a pressure source, we can say it also has
an associated pressure source sample instance, so we can just use it and
always set its data to a PressureState.