observe(): Do not always add a collector to the platform collector mapping
Avoid adding a platform collector with an empty associated pressure source
to the platform collector mapping.

Having observe() attempt to find a suitable platform collector every time in
this case is not that expensive, and _not_ doing it leads to a situation
where inactive observers still need to call unobserve() or disconnect() to
remove an entry from the platform collector mapping (to get observe() to try
to connect to a platform collector again, for example).